Address PR #72 review: consistent counter_trade, trade log presentation, split trade tests
- counter_trade: remove the `state.phase === "action"` gate from
  handleCounterTrade so it behaves like accept/reject — a true global async
  RESPONSE valid off-turn in any phase, matching its registration in
  GLOBAL_ACTION_ROUTES. Propose keeps its action-phase gate (it spends AP);
  countering does not charge AP. Add tests locking counter off-turn / in a
  non-action phase and the no-AP contract.
- gameLogDisplay: add ACTION_PRESENTATION entries for all trade_* log types
  (proposed/accepted/rejected/expired/countered) so they render with proper
  "Trade"/"deal" presentation instead of the generic fallback. Add a test
  asserting non-default presentation.
- Split tests/unit/tradeActions.test.ts (~1k lines) into focused files
  (core, expiry, counter) sharing a tradeActionsFixture helper; no coverage
  loss.
- PhaseActionHandler: keep the uniform nowMs signature (zero churn — non-trade
  closures simply omit the trailing param) and document that only trade
  handlers read nowMs.
